+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Testing with CaptionTrackList plugin
 - Added option to invertColors (`excludeLegends`) to exclude legend item swatch colors
 - Transferred repo to @sciguy [2025-06-03]
+- add highlightRegion to contigs
 
 
 ## 1.7.1 - 2024-10-22